EZ-1301 migrate to exposed 0.25.1 (update build.gradle and use type long in the corresponding places). Since limit still uses type int, a local conversion was done in the function call

+++ core/src/main/kotlin/core/ems/service/exercise/TeacherReadExercises.kt
@@ -62,7 +62,7 @@ private fun selectTeacherExercisesOnCourse(courseId: Long, callerId: String, cou
         val restrictedGroups = getTeacherRestrictedGroups(courseId, callerId)
         val studentQuery = courseService.selectStudentsOnCourseQuery(courseId, emptyList(), restrictedGroups)
 
-        val studentCount = studentQuery.count()
+        val studentCount = studentQuery.count().toInt()
         val students = studentQuery.map { it[Student.id].value }
 
         (CourseExercise innerJoin Exercise innerJoin ExerciseVer)
